requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
catalogue==2.0.10
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
diffusers==0.27.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
g2pkk==0.1.2
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
grpcio==1.63.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.23.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==7.1.0
inflect==7.2.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
marisa-trie==1.1.1
Markdown==3.6
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.8.4
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_311/mkl_random_1676481991689/work
mkl-service==2.4.0
more-itertools==10.2.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
openai-whisper==20231117
packaging==24.0
pandas==1.5.3
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
platformdirs==4.2.1
pooch==1.8.1
preshed==3.0.9
protobuf==5.26.1
psutil==5.9.8
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
regex==2024.5.10
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
trainer==0.0.36
transformers==4.40.2
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3 @ file:///C:/b/abs_4etpfrkumr/croot/urllib3_1707770616184/work
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
yarl==1.9.4
zipp==3.18.1
